backend/pyproject.toml
[tool.poetry]
name = "movie-autotranslate-backend"
version = "0.1.0"
description = "FastAPI backend for the AI dubbing pipeline."
authors = ["Movie Autotranslate Team <devs@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.19"
pydantic-settings = "^2.2.1"
redis = "^5.0.4"
celery = "^5.3.6"
python-multipart = "^0.0.9"
minio = "^7.2.9"
tenacity = "^8.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
ruff = "^0.4.2"
black = "^24.4.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
